+++ Changelog.md
@@ -8,10 +8,10 @@ suggestions, ideas and improvements to FriendlyId.
 
 ## 4.0.3 (NOT RELEASED YET)
 
-* Fix escape for '%' and '_' on SQLite (Norman Clarke and [spetrunin](https://github.com/spetrunin))
+* Fix escape for '%' and '_' on SQLite (Norman Clarke and Sergey Petrunin)
 * Allow FriendlyId to be extended or included (Norman Clarke)
 * Allow Configuration#use to accept a module (Norman Clarke)
-* Fix bugs with History module + STI (Norman Clarke and [spetrunin](https://github.com/spetrunin))
+* Fix bugs with History module + STI (Norman Clarke and Sergey Petrunin)
 
 ## 4.0.2 (2012-03-12)